Move DebouncingQueryTextListener to a separate file
