+++ app/src/main/java/leo/me/la/movies/util/DebouncingQueryTextListener.kt
@@ -1,0 +1,46 @@
+package leo.me.la.movies.util
+
+import androidx.appcompat.widget.SearchView
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.OnLifecycleEvent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+
+
+internal class DebouncingQueryTextListener(
+    lifecycle: Lifecycle,
+    private val onDebouncingQueryTextChange: (String?) -> Unit
+) : SearchView.OnQueryTextListener, LifecycleObserver {
+
+    private val coroutineScope: CoroutineScope = CoroutineScope(Dispatchers.Main)
+
+    private var searchJob: Job? = null
+
+    init {
+        lifecycle.addObserver(this)
+    }
+
+    override fun onQueryTextSubmit(query: String?): Boolean {
+        return false
+    }
+
+    override fun onQueryTextChange(newText: String?): Boolean {
+        searchJob?.cancel()
+        searchJob = coroutineScope.launch {
+            newText?.let {
+                delay(500)
+                onDebouncingQueryTextChange(newText)
+            }
+        }
+        return false
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
+    private fun destroy() {
+        searchJob?.cancel()
+    }
+}
